requirements/jenkins.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file requirements/jenkins.txt requirements/base.txt requirements/testing.in requirements/jenkins.in
#
-e git+git://github.com/GemeenteUtrecht/vng-api-common@feature/drc-cmis#egg=vng_api_common
alabaster==0.7.12
astroid==1.6.5            # via pylint
babel==2.6.0
beautifulsoup4==4.6.0     # via webtest
certifi==2018.4.16
chardet==3.0.4
click==6.7
cmislib-maykin==0.7.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
cssselect==1.0.3          # via pyquery
django-appconf==1.0.2
django-axes==4.4.0
django-choices==1.7.0
django-cors-middleware==1.3.1
django-extra-fields==1.0.0
django-filter==2.0.0
django-ipware==2.1.0
django-jenkins==0.110.0
django-markup==1.3
django-privates==1.0.1.post0
django-sendfile2==0.4.2
django-solo==1.1.3
django-webtest==1.9.3
django==2.2.2
djangorestframework-camel-case==0.2.0
djangorestframework-filters==0.10.2
djangorestframework-gis==0.14
djangorestframework==3.9.4
docutils==0.14
drc-cmis==0.3.2
drf-nested-routers==0.90.2
drf-yasg==1.16.0
factory-boy==2.11.1
faker==0.8.17             # via factory-boy
first==2.0.1
freezegun==0.3.10
gemma-zds-client==0.11.0
httplib2==0.12.1
humanize==0.5.1
idna==2.7
imagesize==1.1.0
import-class==0.1.1
inflection==0.3.1
iso-639==0.4.5
iso-8601==0.3.0
iso8601==0.1.12
isodate==0.6.0
isort==4.3.4
itypes==1.1.0
jinja2==2.10.1
lazy-object-proxy==1.3.1  # via astroid
lxml==4.2.3               # via pyquery
markdown==3.1
markupsafe==1.0
mccabe==0.6.1             # via pylint
oyaml==0.7
packaging==19.0
pep8==1.7.1
pillow==5.2.0
pip-tools==2.0.2
psycopg2-binary==2.7.5
pygments==2.3.1
pyjwt==1.6.4
pylint==1.9.2
pyparsing==2.3.1
pyquery==1.4.0
python-dateutil==2.7.3
python-dotenv==0.8.2
pytz==2018.4
pyyaml==3.12
raven==6.9.0
requests==2.21
ruamel.yaml==0.15.40
six==1.11.0
snowballstemmer==1.2.1
sphinx-rtd-theme==0.4.2
sphinx==1.8.4
sphinxcontrib-websupport==1.1.0
sqlparse==0.3.0
tblib==1.3.2
text-unidecode==1.2       # via faker
unidecode==1.0.22
uritemplate==3.0.0
urllib3==1.24.3
waitress==1.1.0           # via webtest
webob==1.8.2              # via webtest
webtest==2.0.30
wrapt==1.10.11            # via astroid
